+++ server/Simulation/Systems/DropSystem.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Numerics;
 using server.ECS;
 using server.Simulation.Components;
 using server.Simulation.Database;
@@ -17,14 +18,14 @@ public sealed class DropSystem : NttSystem<DeathTagComponent, Box2DBodyComponent
         // get random Db.BaseResource
         var randomId = Random.Shared.Next(3, 7);
         var resource = Db.BaseResources[randomId];
-
+        var killerPos = dtc.Killer.Get<Box2DBodyComponent>().Position;
 
         for (var i = 0; i < pik.Amount; i++)
         {
             var lifetime = TimeSpan.FromMilliseconds(Random.Shared.Next(5000, 10000));
-            var direction = SpawnManager.GetRandomDirection();
+            var direction = Vector2.Normalize(killerPos - rigidBody.Position);
             var position = rigidBody.Position + (direction * 2);
-            SpawnManager.SpawnDrop(Db.BaseResources[Random.Shared.Next(3, 8)], position, lifetime, direction * 100);
+            SpawnManager.SpawnDrop(Db.BaseResources[Random.Shared.Next(3, 8)], position, lifetime, direction * 10f);
         }
     }
 }